[Feat] 함수 일부 분리
